Implemented party finder notifications
This commit includes several significant changes. The namespace of `Job.cs` was changed. `Slot.cs` was modified to a class with two properties. A new dictionary was added in `SubscribeToPf.cs` to map shorthand duty names to full names, and several methods were updated or replaced. A new command `pfalert` was added to subscribe to notifications. `Program.cs` now includes a new singleton service. In `PartyFinderService.cs`, the `UpdatePartyFinderListings` method was replaced with `HandleUpdate` method and a new method `SendNotifications` was added. A new record `PFNotificationRequest` was added to represent a notification request, and a new class `PartyMatcher` was added to match party compositions.